Revert "Merge pull request #1373 from square/zachklipp/snapshotparcels"
This reverts commit 9410118d3209e8723c2a042c0a464836788087cd, reversing
changes made to fc9e379e868b65e06db9bcff7d7fd70e9f8784a5.
