--- workflow-core/dependencies/androidRuntimeClasspath.txt
@@ -1,11 +1,0 @@
-com.squareup.okio:okio-jvm:3.3.0
-com.squareup.okio:okio:3.3.0
-org.jetbrains.kotlin:kotlin-bom:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib-common:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib-jdk7:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib-jdk8:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib:2.0.21
-org.jetbrains.kotlinx:kotlinx-coroutines-bom:1.7.3
-org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.7.3
-org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3
-org.jetbrains:annotations:23.0.0
