--- workflow-core/src/jvmMain/kotlin/com/squareup/workflow1/CommonUniqueClassName.jvm.kt
@@ -1,7 +1,0 @@
-package com.squareup.workflow1
-
-import kotlin.reflect.KClass
-
-internal actual fun commonUniqueClassName(kClass: KClass<*>): String {
-  return kClass.qualifiedName ?: kClass.toString()
-}
