+++ workflow-core/src/jvmMain/kotlin/com/squareup/workflow1/CommonUniqueClassName.kt
