+++ workflow-ui/core-android/src/main/java/com/squareup/workflow1/ui/SnapshotParcels.kt
@@ -1,9 +1,10 @@
 package com.squareup.workflow1.ui
 
+import android.os.Build.VERSION
+import android.os.Build.VERSION_CODES
+import android.os.Parcel
 import android.os.Parcelable
 import com.squareup.workflow1.Snapshot
-import com.squareup.workflow1.toParcelable
-import com.squareup.workflow1.toSnapshot
 import okio.ByteString
 
 /**
@@ -12,23 +13,35 @@ import okio.ByteString
  *
  * Read the [Parcelable] back with [toParcelable].
  */
-@Deprecated(
-  "Use toSnapshot() from workflow-core instead.",
-  replaceWith = ReplaceWith("toSnapshot()", "com.squareup.workflow1.toSnapshot")
-)
-public fun Parcelable.toSnapshot(): Snapshot = toSnapshot()
+public fun Parcelable.toSnapshot(): Snapshot {
+  return Snapshot.write { bufferedSink ->
+    val parcel = Parcel.obtain()
+    parcel.writeParcelable(this, 0)
+    val byteArray = parcel.marshall()
+    bufferedSink.write(byteArray)
+    parcel.recycle()
+  }
+}
 
 /**
  * @return a [Parcelable] previously wrapped with [toSnapshot], or `null` if the receiver is empty.
  */
-@Deprecated(
-  "Use toParcelable() from workflow-core instead.",
-  replaceWith = ReplaceWith("toParcelable()", "com.squareup.workflow1.toParcelable")
-)
-public inline fun <reified T : Parcelable> Snapshot.toParcelable(): T? = toParcelable()
+public inline fun <reified T : Parcelable> Snapshot.toParcelable(): T? {
+  return bytes.takeIf { it.size > 0 }
+    ?.toParcelable<T>()
+}
 
-@Deprecated(
-  "Use toParcelable() from workflow-core instead.",
-  replaceWith = ReplaceWith("toParcelable()", "com.squareup.workflow1.toParcelable")
-)
-public inline fun <reified T : Parcelable> ByteString.toParcelable(): T = toParcelable<T>()!!
+public inline fun <reified T : Parcelable> ByteString.toParcelable(): T {
+  val parcel = Parcel.obtain()
+  val byteArray = toByteArray()
+  parcel.unmarshall(byteArray, 0, byteArray.size)
+  parcel.setDataPosition(0)
+  val rtn = if (VERSION.SDK_INT >= VERSION_CODES.TIRAMISU) {
+    parcel.readParcelable<T>(Snapshot::class.java.classLoader, T::class.java)!!
+  } else {
+    @Suppress("DEPRECATION")
+    parcel.readParcelable<T>(Snapshot::class.java.classLoader)!!
+  }
+  parcel.recycle()
+  return rtn
+}
